cms-admin: Move TipTapContentTranslationDialog into its own file
createTipTapRichTextBlock.tsx keeps growing with each translation-related
addition. Move the review-dialog wrapper component into its own file,
exporting TipTapEditor and TipTapEditorProps for it to reuse, matching the
one-component-per-file convention already applied to the translation logic
and stories.
